+++ CHANGELOG.md
@@ -1,12 +1,24 @@
 # Changelog
 
-## 1.1.0 / 2018-01-03
+## Ongoing [☰](https://github.com/philnash/jekyll-gzip/compare/v1.1.0...master)
+
+### Added
+
+* Hooks into Jekyll::Assets if available
+
+### Changed
+
+* Uses built in `Jekyll.env` instead of `ENV["JEKYLL_ENV"]`
+* Changes `Jekyll::Zopfli::Compressor` to a module and implements a `compress_directory` method
+* Moves Jekyll::Zopfli::ZIPPABLE_EXTENSIONS into plugin config that can overwritten in the site config
+
+## 1.1.0 (2018-01-03) [☰](https://github.com/philnash/jekyll-zopfli/compare/v1.0.0...v1.1.0)
 
 ### Changed
 
 * Only run the post write hook when the environment variable `JEKYLL_ENV` is `production`
 
-## 1.0.0 / 2018-01-02
+## 1.0.0 (2018-01-02) [☰](https://github.com/philnash/jekyll-zopfli/commits/v1.0.0)
 
 ### Added